requirements.txt
peft>=0.1.0
requests
jieba
numpy
networkx
#llama-cpp-python
transformers
Flask
bs4
flask_cors
langchain
sentencepiece
cpm_kernels
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
chromadb